refactor(firestore): extract atomic batch execution
